[#5161] Storage - ContentAccessApi fix

+++ api/src/test/java/org/smartbit4all/api/contentaccess/ContentAccessTest.java
@@ -35,8 +35,10 @@ public class ContentAccessTest {
   @BeforeAll
   static void clearContentAccessRootFolder() {
     File root = new File(UPLOAD_FOLDER_PATH);
-    for (File innerFile : root.listFiles()) {
-      innerFile.delete();
+    if (root.exists()) {
+      for (File innerFile : root.listFiles()) {
+        innerFile.delete();
+      }
     }
   }
